Move transition class to subproject
