+++ fabric/src/main/java/software/bernie/geckolib/network/packet/EntityAnimTriggerPacket.java
@@ -7,14 +7,13 @@ import net.minecraft.client.multiplayer.ClientPacketListener;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.Entity;
-import software.bernie.TransitionRenderUtils;
+import software.bernie.geckolib.TransitionRenderUtils;
 import software.bernie.geckolib.animatable.GeoEntity;
 import software.bernie.geckolib.animatable.GeoReplacedEntity;
 import software.bernie.geckolib.core.animatable.GeoAnimatable;
 import software.bernie.geckolib.network.AbstractPacket;
 import software.bernie.geckolib.network.GeckoLibNetwork;
 import software.bernie.geckolib.util.ClientUtils;
-import software.bernie.geckolib.util.RenderUtils;
 
 import javax.annotation.Nullable;
